fix(#2): add access modifiers | set LocalDate instead of LocalDateTime | create 2 new enums | fix domain package naming
