+++ src/main/java/com/velocity/api/bike/BikeCategory.java
@@ -1,0 +1,5 @@
+package com.velocity.api.bike;
+
+public enum BikeCategory {
+    CARGO
+}
