+++ src/main/java/com/velocity/api/bike/BikeStatus.java
@@ -1,4 +1,4 @@
-package com.velocity.velocity.api.bike;
+package com.velocity.api.bike;
 
 public enum BikeStatus {
     ACTIVE, MAINTENANCE, LOST, RETIRED
