+++ src/main/java/com/velocity/api/bike/dto/BikeInstanceDto.java
@@ -1,7 +1,5 @@
-package com.velocity.velocity.api.bike.dto;
+package com.velocity.api.bike.dto;
 
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
 import lombok.Setter;
