+++ src/main/java/com/velocity/api/reservation/ReservationStatus.java
@@ -1,0 +1,5 @@
+package com.velocity.api.reservation;
+
+public enum ReservationStatus {
+    PENDING, CONFIRMED, COMPLETED, CANCELLED
+}
