+++ src/main/java/com/velocity/api/reservation/dto/ReservationDto.java
@@ -1,13 +1,18 @@
-package com.velocity.velocity.api.reservation.dto;
+package com.velocity.api.reservation.dto;
+
+import lombok.Getter;
+import lombok.Setter;
 
 import java.math.BigDecimal;
-import java.time.LocalDateTime;
+import java.time.LocalDate;
 import java.util.UUID;
 
+@Getter
+@Setter
 public class ReservationDto {
     private UUID id;
-    private LocalDateTime startDate;
-    private LocalDateTime endDate;
+    private LocalDate startDate;
+    private LocalDate endDate;
     private BigDecimal totalCost;
     private String status;
     private UUID userId;
