+++ src/main/java/com/velocity/api/user/UserCity.java
@@ -1,0 +1,5 @@
+package com.velocity.api.user;
+
+public enum UserCity {
+    GDANSK, WROCLAW, WARSAW, POZNAN
+}
