+++ src/main/java/com/velocity/api/user/UserStatus.java
@@ -1,4 +1,4 @@
-package com.velocity.velocity.api.user;
+package com.velocity.api.user;
 
 public enum UserStatus {
     ACTIVE, BLOCKED, DELETED
