+++ src/main/java/com/velocity/api/user/dto/UserDto.java
@@ -1,7 +1,5 @@
-package com.velocity.velocity.api.user.dto;
+package com.velocity.api.user.dto;
 
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
 import lombok.Setter;
