--- src/main/java/com/velocity/velocity/api/bike/dto/BikeModelDto.java
@@ -1,13 +1,0 @@
-package com.velocity.velocity.api.bike.dto;
-
-import java.math.BigDecimal;
-import java.util.UUID;
-
-public class BikeModelDto {
-    private UUID id;
-    private String name;
-    private String description;
-    private BigDecimal speed;
-    private int range;
-    private int capacity;
-}
