--- src/main/java/com/velocity/velocity/api/reservation/ReservationStatus.java
@@ -1,5 +1,0 @@
-package com.velocity.velocity.api.reservation;
-
-public enum ReservationStatus {
-    PENDING, ACTIVE, COMPLETED, CANCELLED
-}
