--- src/main/java/com/velocity/velocity/api/user/UserCity.java
@@ -1,5 +1,0 @@
-package com.velocity.velocity.api.user;
-
-public enum UserCity {
-    GDANSK, KRAKOW, WARSZAWA, POZNAN
-}
