+++ src/test/java/com/velocity/api/VelocityApiApplicationTests.java
@@ -1,4 +1,4 @@
-package com.velocity.velocity.api;
+package com.velocity.api;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
